var_info_distil/requirements.txt
torch==1.3.1
torchvision==0.4.2
pytorch-ignite==0.3.0